Switch to using int64 Convex type for duration
This requires android-convexmobile:0.4.0, which has fixes for properly serializing Int, Long, Float and Double for Convex.

Also makes a small tweak to allow the Workout.duration to be missing (previously the app would crash).

+++ convex/schema.ts
@@ -5,7 +5,7 @@ export default defineSchema({
   workouts: defineTable({
     activity: v.string(),
     date: v.string(), // Must be in format "YYYY-MM-DD" a.k.a ISO 8601 format
-    duration: v.optional(v.float64()),
+    duration: v.optional(v.int64()),
     userId: v.string(),
   }).index("userId_date", ["userId", "date"]),
 });
